snprintf format argument

+++ src/impl/tx.cpp
@@ -9,6 +9,7 @@
 #include <cassert>
 #include <chrono>
 #include <mutex>
+#include <inttypes.h>
 
 namespace bolt::impl {
 
@@ -333,7 +334,7 @@ std::future<std::vector<std::string>> Tx::Check() {
       for (auto item : all) {
         auto it = freed.find(item);
         if (it != freed.end()) {
-          snprintf(buf, sizeof(buf), "page %lld: already freed", item);
+          snprintf(buf, sizeof(buf), "page %" PRIu64 ": already freed", item);
           errors.push_back(buf);
         }
         freed[item] = true;
@@ -354,7 +355,7 @@ std::future<std::vector<std::string>> Tx::Check() {
         auto it = reachable.find(i);
         auto itf = freed.find(i);
         if (it == reachable.end() && itf == freed.end()) {
-          snprintf(buf, sizeof(buf), "page %lld: unreachable unfreed", i);
+          snprintf(buf, sizeof(buf), "page %" PRIu64 ": unreachable unfreed", i);
         }
       }
       return errors;
@@ -379,8 +380,8 @@ void Tx::checkBucket(impl::BucketPtr bucket,
     txptr->forEachPage(bucket->bucket.root, 0, [&](impl::page *p, int depth) {
         char buf[1024];
         if (p->id > txptr->meta.pgid) {
-            snprintf(buf, sizeof(buf), "page %d: out of bounds: %d", (int)p->id,
-                     (int)txptr->meta.pgid);
+            snprintf(buf, sizeof(buf), "page %" PRIu64 ": out of bounds: %" PRIu64, p->id,
+                     txptr->meta.pgid);
             errors.push_back(buf);
         }
         // Ensure each page is only referenced once.
@@ -388,7 +389,7 @@ void Tx::checkBucket(impl::BucketPtr bucket,
             auto id = p->id + 1;
             auto it = reachable.find(id);
             if (it != reachable.end()) {
-                snprintf(buf, sizeof(buf), "page %d: multiple references", (int)id);
+                snprintf(buf, sizeof(buf), "page %" PRIu64 ": multiple references", id);
                 errors.push_back(buf);
             }
             reachable[id] = p;
@@ -397,11 +398,11 @@ void Tx::checkBucket(impl::BucketPtr bucket,
         // We should only encounter un-freed leaf and branch pages.
         auto it = freed.find(p->id);
         if (it != freed.end()) {
-            snprintf(buf, sizeof(buf), "page %d: reachable freed", (int)p->id);
+            snprintf(buf, sizeof(buf), "page %" PRIu64 ": reachable freed", p->id);
             errors.push_back(buf);
         } else if ((p->flags & impl::branchPageFlag) == 0 &&
                    (p->flags & impl::leafPageFlag) == 0) {
-            snprintf(buf, sizeof(buf), "page %d: invalid type: %s", (int)p->id,
+            snprintf(buf, sizeof(buf), "page %" PRIu64 ": invalid type: %s", p->id,
                      p->type().c_str());
             errors.push_back(buf);
         }
